modification des vitesse de capteur.cpp et de intersectionGauche

+++ leGeant.com/librairie/librairieFonctions.cpp
@@ -65,7 +65,7 @@ void initialisationPwmMoteurs(){                        //TIMER 0 (B3 et B4)
     sei();
 }
 
-void ajustementPwmMoteurs(uint8_t pourcentageA, uint8_t pourcentageB) {     //TIMER 0 (B3 et B4)
+void ajustementPwmMoteurs(float pourcentageA, float pourcentageB) {     //TIMER 0 (B3 et B4)
 
     pourcentageA *= 0.85;       //Coefficient de vitesse de la roue gauche (ajustement, afin que les roues tournent à la même vitesse).
     OCR1A = 255 * (100.0 - pourcentageA)/100.0;
